refactor tool_defs.py to separate concerns, fix ruff to keep python 3.9 compat
